+++ src/main/scala/org/constellation/checkpoint/CheckpointBlockValidator.scala
@@ -6,10 +6,10 @@ import cats.implicits._
 import io.chrisdavenport.log4cats.SelfAwareStructuredLogger
 import io.chrisdavenport.log4cats.slf4j.Slf4jLogger
 import org.constellation.DAO
+import org.constellation.domain.transaction.TransactionValidator
 import org.constellation.primitives.Schema.CheckpointCache
 import org.constellation.primitives.{CheckpointBlock, Transaction}
 import org.constellation.storage.{AddressService, SnapshotService}
-import org.constellation.transaction.TransactionValidator
 import org.constellation.util.{HashSignature, Metrics}
 
 object CheckpointBlockValidator {
