+++ src/test/scala/org/constellation/storage/transactions/TransactionGossipingTest.scala
@@ -6,7 +6,7 @@ import cats.implicits._
 import io.chrisdavenport.log4cats.{Logger, SelfAwareStructuredLogger}
 import io.chrisdavenport.log4cats.slf4j.Slf4jLogger
 import org.constellation.domain.consensus.ConsensusStatus
-import org.constellation.domain.transaction.{TransactionChainService, TransactionService}
+import org.constellation.domain.transaction.{TransactionChainService, TransactionGossiping, TransactionService}
 import org.constellation.p2p.PeerData
 import org.constellation.primitives.{Transaction, TransactionCacheData}
 import org.constellation.{ConstellationExecutionContext, DAO, Fixtures}
