+++ src/test/scala/org/constellation/tx/TXValidationBenchmark.scala
@@ -4,7 +4,7 @@ import com.typesafe.scalalogging.Logger
 import org.constellation.Fixtures._
 import org.constellation.crypto.KeyUtils
 import org.constellation.crypto.KeyUtils._
-import org.constellation.transaction.TransactionValidator
+import org.constellation.domain.transaction.TransactionValidator
 import org.scalatest.FlatSpec
 
 class TXValidationBenchmark extends FlatSpec {
